moved classes to new package to improve modularization
