+++ server/src/main/java/cn/keking/config/configconstants/MediaConfigConstants.java
@@ -1,4 +1,4 @@
-package cn.keking.config;
+package cn.keking.config.configconstants;
 
 import org.springframework.beans.factory.annotation.Value;
